Back-end/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.11
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
importlib_metadata==7.1.0
Markdown==3.6
PyJWT==2.8.0
requests==2.31.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
zipp==3.18.1
